Начало работы, доработал главный экран
