+++ package.json
@@ -10,7 +10,10 @@
     "start": "vite build && vite preview"
   },
   "devDependencies": {
+    "@types/handlebars": "^4.0.40",
+    "@types/node": "^24.3.1",
     "sass": "^1.91.0",
+    "typescript": "^5.9.2",
     "vite": "^7.1.4",
     "vite-plugin-handlebars": "^2.0.0"
   },
