--- src/components/Input.js
@@ -1,4 +1,0 @@
-export default `
-<input id="{{id}}" type="{{type}}" class="{{class}}" placeholder="{{placeholder}}" name="{{name}}" required>
-`; 
-
