+++ src/components/Input.ts
@@ -1,0 +1,4 @@
+export default `
+<input id="{{id}}" type="{{type}}" class="{{class}}" placeholder="{{placeholder}}" name="{{name}}" {{required}}>
+`; 
+
