+++ src/pages/404/index.ts
