+++ src/pages/500/index.ts
