+++ src/pages/allpages/index.ts
