+++ src/pages/login/login.hbs
@@ -2,8 +2,8 @@
     <form class="auth-form">
         <h1 class="auth-form__title">Вход</h1>
 
-        {{> Input id="login-username" class="auth-form__input" type="login" placeholder="Логин" name="login"}}
-        {{> Input id="login-password" class="auth-form__input" type="password" placeholder="Пароль" name="password"}}
+        {{> Input id="login-username" class="auth-form__input" type="login" placeholder="Логин" name="login" required="required"}}
+        {{> Input id="login-password" class="auth-form__input" type="password" placeholder="Пароль" name="password" required="required"}}
         
         {{> Button id="submit-btn"  class="auth-form__btn login-btn" text="Авторизироваться"}}
