+++ src/styles/_mixins.scss
@@ -15,4 +15,3 @@
     background: darken($bg-color, 10%);
   }
 }
-
