+++ src/styles/main.scss
@@ -17,12 +17,24 @@ input, textarea, select, button {
     margin: 0px;
 }
 
+// Убираем вертикальный скролл
+
+body, html {
+    height: 100%;
+    overflow: hidden;
+}
+
 .button, a { 
     cursor: pointer;
 }
 
 a {
     cursor: pointer;
+}
+
+ul {
+    list-style-type: none;
+    padding: 0;
 }
 
 .container {
@@ -35,6 +47,10 @@ a {
 
 .flex-wrapper {
     display: flex;
+}
+
+.hide {
+    display: none !important;
 }
 
 // Pages: login, register
@@ -124,7 +140,6 @@ a {
     &-nav {
         display: flex;
         flex-direction: column;
-        border-bottom: 1px solid $bg-white-border;
         padding: 20px 10px 15px 10px;
 
         &__profile-link {
@@ -150,15 +165,116 @@ a {
     }
 }
 
-.chats-picked {
-    width: 75%;
-    height: 100vh;
+.chat-card {
     display: flex;
-    justify-content: center;
-    align-items: center;    
-    height: 100vh;          
-    width: 100vw;           
+    padding: 12px 10px;
+    cursor: pointer;
+    border-top: 1px solid $bg-white-border;
+    border-color: $bg-white-border;
+
+    &__avatar-wrapper {
+        margin-right: 3%;
+        &_image {
+            max-width: 72px;
+            max-height: 72px;
+        }
+    }
+
+    &__text {
+        display: flex;
+        flex-direction: column;
+
+        &_name {
+            font-weight: 600;
+            font-size: 13px;
+            display: block;
+            margin-bottom: 6px;
+        }
+
+        &_last-message {
+            flex-wrap: wrap;
+            display: block;
+            font-size: 12px;
+            line-height: 15px;
+        }
+    }
+
+    &__numbers {
+        display: flex;
+        flex-direction: column;
+        margin-left: auto;
+        justify-content: space-between;
+
+        &_message-time {
+            font-weight: 500;
+            font-size: 9px;
+            line-height: 13px;
+            color: #999;
+        }
+
+        &_unread-count {
+            font-weight: 500;
+            font-size: 11px;
+            color: #FFF;
+            padding: 4px 8px;
+            border-radius: 16px;
+            background-color: $primary-color;
+        }
+    }
 }
+
+
+
+.picked-chat {
+    display: flex;
+    flex-direction: column;
+
+    // Заглушка
+    &__nothing-picked {
+        display: flex;
+        justify-content: center;
+        align-items: center;    
+        height: 94%;          
+        width: 75vw;        
+    }
+
+    // Форма
+    &__message-form {
+            display: flex;
+            justify-content: space-between;
+            height: 6%;
+            background-color: #f5f5f5;
+            border-top: 1px solid #eaeaea;
+            padding: 10px 20px;
+            align-items: center;
+        &_file-img {
+            cursor: pointer;
+            margin-right: 1%;
+        }
+
+        &_file-input-hidden {
+            position: absolute;
+            width: 0;
+            height: 0;
+            opacity: 0;
+            overflow: hidden;
+        }
+
+        &_input {
+            width: 100%;  
+            padding: 10px 20px;
+            border: 1px solid #eaeaea;
+            border-radius: 20px;
+            outline: none;
+        }
+
+        &_send-message-btn {
+            cursor: pointer;
+            margin-left: 1%;
+        }
+    }
+}
+
 
 // Pages: profile, edit-profile, edit-password
 
@@ -263,11 +379,6 @@ a {
 
 .cheatpages-list-wrapper {
     margin: 0 auto;
-}
-
-.cheatpages-list {
-    list-style-type: none;
-    padding: 0;
 }
 
 .cheatpages-list__page_link {
