+++ src/types/handlebars.d.ts
@@ -1,0 +1,4 @@
+declare module '*.hbs?raw' {
+  const content: string;
+  export default content;
+}
